requirements.txt
pyusb
pillow
pycurl
certifi
argparse
netifaces
pluggy